refactor(gfql/polars): move degree helpers out of chain.py into degrees.py
get_degrees_polars / _single_direction_degree_polars / get_indegrees_polars /
get_outdegrees_polars are node-degree computations, not chain/traversal orchestration —
extracted from chain.py into a dedicated lazy/engine/polars/degrees.py. The generic
frame-shape helpers _is_lazy/_colnames (used by chain 8x + the degree fns) move DOWN into
the shared dtypes.py as is_lazy/colnames, so both chain and degrees import them (no
helper->orchestration back-dependency). Only importer (gfql/call/executor.py) repointed;
tests reach the fns via .get_degrees()/.get_indegrees()/.get_outdegrees() dispatch, unchanged.
Behavior-identical; degree + polars chain traversal smoke-verified locally.
