+++ graphistry/compute/gfql/call/executor.py
@@ -76,13 +76,13 @@ def _execute_validated_call(g: Plottable, function: str, validated_params: Dict[
     # declined by that guard.
     if _active_frames_are_polars(g):
         if function == "get_degrees":
-            from graphistry.compute.gfql.lazy.engine.polars.chain import get_degrees_polars
+            from graphistry.compute.gfql.lazy.engine.polars.degrees import get_degrees_polars
             return get_degrees_polars(g, **validated_params)
         if function == "get_indegrees":
-            from graphistry.compute.gfql.lazy.engine.polars.chain import get_indegrees_polars
+            from graphistry.compute.gfql.lazy.engine.polars.degrees import get_indegrees_polars
             return get_indegrees_polars(g, **validated_params)
         if function == "get_outdegrees":
-            from graphistry.compute.gfql.lazy.engine.polars.chain import get_outdegrees_polars
+            from graphistry.compute.gfql.lazy.engine.polars.degrees import get_outdegrees_polars
             return get_outdegrees_polars(g, **validated_params)
 
     if not hasattr(g, function):
